Allow Orientation pagedevice parameter to control the FitPage rotation
With this change, Orientation is no longer ignored except for devices
that support pdfmark where it must be 0 (e.g., pdfwrite). Now with
-dFitPage, -c "<< /Orientation 0 >> setpagedevice" will use either 0
or 90 degree rotation, while -c "<< /Orientation 2 >> setpagedevice"
will rotate either 180 degrees or 270 degrees.

Note that Orientation 2 will invert a page (rotate by 180) if 270
degree rotation is not needed for the best fit.

Orientation will also control which way is "up" without -dFitPage,
but may result in clipping.