Fix 'warning: ISO C90 forbids mixed declarations and code'
Commit f02d0a5 added code which declares variables in the body of a
function. This is not permitted in C90 (or C89), is contrary to the coding
style for Ghostscript and fails to compile Ghostscript under Visual Studio.

In addition 'zip_close_all_archive_files()' did not have a return value on
success which led to random return values and a number of regression files
failing.

Moved the variables to the start of the functions (removing const as required)
and added a specific return code on success.

Now compiles under Windows, does not throw as many warnings on Linux and
does not cause spurious regression failures. There are still many warnings
produced by gcc, EG:

./devices/vector/gdevxps.c:1771:9: warning: unused variable 'count' [-Wunused-variable]
./devices/vector/gdevxps.c:1770:9: warning: unused variable 'code' [-Wunused-variable]

The author of the commit should address those separately.