NTFS - fix some memory problems Bug 695752
Bug #695752 "gp_enumerate_files_close leaks some memory"

This appears to be a simple oversight in the original commit, a piece of
allocated memory was not freed, the 'free' call was commented out.

Noticed in passing; if the allocation of the file pattern fails, we were not
freeing the file or directory enumerators, leading to a memory leak as
well.

No differences expected (cluster does not test Windows file system)