txtwrite - properly terminate 'char' tags
Bug #695788 "txtwrite textoutput=0 XML file"

The 'char' tag was being terminated with a '>' instead of a '/>'.

Reported, and the offending code isolated, by Petko Lukov, who offered to
create a patch, but the change is so trivial I simply went ahead with it.

No differences expected, cluster doesn't test txtwrite.