Bug 695774: Fallback for glyph with degenerate bbox
(Re-applying after update to Freetype 2.5.5)

If a glyph has only one very narrow feature, the bbox can end up with either
the width or height of the bbox being 0, in which case no raster memory is
allocated and no attempt is made to render the glyph. This is less than ideal
when the drop-out compensation in the rendering code would actually result in
the glyph being rendered.

Normally the coordinates for the bbox are rounded (which is technically
correct). Now, if that results in the width or height being 0, we'll
then try explicitly rounding up/down, providing a sort of naive whole glyph
drop-out compensation.

The same change has been included upstream and will be in the next scheduled
Freetype release:
http://git.savannah.gnu.org/cgit/freetype/freetype2.git/commit/?id=942aa5ff

This causes a large number of low resolution differences on cluster, all
progressions where glyphs like dash ('-'), underscore ('_') and vertical
bar ('|') were dropping out entirely, and are now printing.