update the mswinpr2 device for FirstPage and LastPage
The device keeps a copy of its original pointer in its structure, ostensibly
to check if it is a copy made by copydevice. If we don't update this pointer
after subclassing,the device doesn't work at all, it produces no output.

This of course suggests that whatever the original rationale behind keeping
the copy, it doesn't actually work!

Here we just update the pointer, I'm not prepared to go delving any further
through this device to try and figure out what the pointer is supposed to do.