Noticed by chrisl while working on moving our installed base fonts to CFF, if a type 0 font has an FDepVector which uses both a type 3 font, and a base font (which is CFF) then we fail to process the condition where we use one glyph from the type 3 and follow it with a glyph from the type 2.
The problem seems to be that we get confused, we detect the error with the
CFF font, but instead of capturing it as a bitmap (fallback case) we end
up dealing with it as if we had started capturing a type 3 charproc. Of
course we didn't start capturing it that way, so that causes chaos.

In the long run this will need to be fixed properly (ie figure out why we
are ending the glyph with the current font being type 3, when it should be
type 2), but for now this prevents the GPF. When/if we do move to CFF fonts
for our base fonts, this will need to be revisited.