Add code to use the new PCL mono palette device when swapping to monochrome mode, instead of hacking the device's color_mapping_procs.
Carries a #if for now for easier debugging.

Currently this code does *NOT* work with pdfwrite and friends, because
many of the device structures maintained by pdfwrite keep a pointer to the
pdfwrite device, which fails badly when devices are added and removed.

pdfwrite structures need to be modified so as not to maintain a pointer to
the device.