Improve high level devices handling of Forms
Bug #695805 "PostScript can defeat the code to spot PostScript forms being run multiple times "

High level devices (eg pdfwrite) already check Forms to see if they are
duplicates, and if so use the initial definition. However, this does not
improve performance, since the Form PaintProc still needs to be executed
for each instance of the Form.

This commit adds two new spec_op enums, and uses them to improve performance
as well, but *only* for high level devices.

After running a PaintProc, the PostScript .execform1 procedure now sends a
spec_op to the device requesting an integer ID to store in the Form. If the
device does not do its own form caching it will return -1, in which case
the PostScript will not store the value.

Each time a Form is executed, the PostScript code will look to see if we
have a non-null value stored for the Implementation key in the Form dictioanry.
If so, then we assume its a dictionary (ideally we should check, but this is
all tightly coupled code), pull the /FormID from it, and then send another
(different) spec_op to the device. On receipt of this the device simply
reuses its stored Form.

This absolves us from running the PaintProc again, which on some files can
have significant performance benefits. It also means that even Forms which
execute pdfmarks (as the original file does) will be recognised as the same
form, which the duplicate removal code cannot achieve.

The customer's test file runs in ~16 seconds now,producing a 4.3 Mb output
file, previously this produced an 81 Mb file and took around 20 minutes.