Fixes for named color replacement.
The sample code for named color replacement had several issues
that this commits fixes.

1) The memory allocations were not being done in non-gc memory.
2) There needed to be a release method to free the allocations
   when we are done with the named color profile structure.
   Since this is a "custom" profile, the release procedure for
   a profile was added as a member variable to the profile structure.
3) The decision to use the named color profile was moved earlier
   in the ghostscript color remapping logic.  The ensures we
   have better control and avoid unwanted multiple remaps.

Once the color is assigned values by the call to
gsicc_transform_named_color in gx_remap_named_color, note
that in gx_remap_concrete_devicen the transfer functions are applied.
Also note that in gx_remap_concrete_devicen we will end up doing
a possible ICC mapping of just the CMYK values from source CMYK to
destination CMYK via the call to devicen_icc_cmyk.  The reason for
this is explained by the large comment in that part of the code.
If this remap is not desired, for example if gx_remap_named_color
assigned the CMYK values in the destination space already, then
this mapping should be disabled. The decision to enable or disable
is a design/customization decision of which the developer should be aware.