Fix for crash issue in -dUseFastColor
The noncm link was storing the target device color mapping procs
and using those to map the colors.  The problem is that some devices
(for example tiffsep) rely upon special device structures to
do the mapping.  When we had a forwarding device like the overprint
device in front of the target, we were calling the target mapping
procs where we should be calling the overprint device mapping procs.
This commit fixes the issue.