Fix error recovery after invalid type 1c (CFF) font
Bug #695819 "Ghostscript cannot process font stream"

The PDF file (which purports to have been produced by an earlier version
of Ghostscript) contains a CFF font which has en empty names index.

Ghostscript, not unreasonably, treats this as being 'the first byte of the
name is 0x00' which indicates a font which has been deleted from the
FontSet. This means that the embedded FontSet contains no fonts.

Since there are on fonts, we obviously can't use the font stream, and so
we enter an error recovery process. This exposes a problem specific to
this type of font, failing in this fashion, which left the operand stack
in a condition where we were unable to recover from the error.

This commit detects the error condition at an earlier stage and prevents
the change to the operand stack, which allows the missing font to be
substituted, and the file runs to completion.

No differences expected.