Fix pattern-clist with transparency color mismatch with raster patterns
Cust 532 noticed that colors would be (very slightly) different when patterns
were rendered from raster patterns vs. using the pattern-clist approach. This
happened when the pattern has transparency such as fts_25_2526.ps.

The file tests_private/comparefiles/Bug692720.pdf is a progression (lower left
"21h" and "22h" are missing and the bold outline for some of the text on the
left panel is wrong without this change.

Fix it to make sure and push and pop a transparency group around the pattern
filling when pattern-clist is used, the same as for raster patterns that have
transparecncy. The 'blending_mode' was moved to the tile (gx_color_tile_s)
from the ttrans (gx_pattern_trans_s) struct so that it will be available when
painting from the pattern-clist.

Also remove some code in gxpcmap.c that was dead (#if 0) since 2007.