Add LockColorants option to tiffsep and psdcmyk devices
If someone specifies a list of colorants using something like this

-c "<< /SeparationColorNames [ /Cyan /Magenta /Yellow /Black /Violet /Orange]
/SeparationOrder [ /Cyan /Magenta /Yellow /Black /Violet /Orange]>> setpagedevice"

on the command line, it would be nice if they also had the option to
keep the colorants to only this set for the tiffsep and psdcmyk devices.

Currently, the list of colorants above would be included as well as any that
are encountered in the document (up to a max of 64).  In some cases, this
may be the desired outcome.  In other cases, the user may like to keep the colorants
to only those set by SeparationColorNames.  Toward that end, this commits adds
a -dLockColorants=true/false option for the tiffsep and psdcmyk devices to
ensure that the devices do not add on additional colorants as they are
encountered in the document.