Use different remap proc for indexed color space when dealing with named colors
Currently the Index color space uses gx_default_remap_color for its remap procedure.
This is a problem for when we are dealing with an index color whose base space
is DeviceN and we are doing named color replacement in our color management.  To
fix this, we introduce a different version of the Indexed color space that is installed
with a different remap proc for the case when we have a index color space whose
base space is DeviceN or Separation AND we have specified a named color ICC profile.