Fix issues with sICCOutputColors option
If an NCLR ICC profile is specified for output then it is also
possible to specify the colorant names associated with the profile.
This information is described by the sICCOutputColors parameter.
There was an issue where the default colorant names AND the
colorant names specified by sICCOutputColors were getting
added to the device.   With this fix only the colorants specified
by sICCOutputColors are added to the device.