Fixes for named color support
A few minor fixes for named color support including the detection of the
None colorant in the DeviceN list of colorants.   Making sure that the
equivalent CMYK color is correctly computed.  Make sure that all the
colorants in the device are cleared out prior to setting the ones
associated with the device profile in the example code given in
gsicc_transform_named_color.