Continuation of moving the clist cache 'chunk' form a static part of the clist device to a dynamically allocation.
It turns out that we cannot free the memory used in clist_close, because the
pattern clist closes the device, but holds on to it in order to replay the
clist later.

So unfortunately we have to free the memory in 2 places, once in
gdev_prn_tear_down() for regular clist devices, and once in
gx_pattern_cache_free_entry() for pattern clists. Kind of ugly but there
seems to be no alternative :-(