Add an image enumerator to the PCL monochrome device. We need to have the rendering code call through our device's color_mapping_procs (which will pass on to the original device's procs).
The graphics library gets the device from the image enumerator, and previously
we were just returning the child enumerator, which meant that the graphics
library was calling the child device's colour mapping procs directly. Here
we add an image enumerator as *well* as creating an enumerator for the child.
Our enumerator simply passes everything on to the child, but it means that
when the graphics library gets the device, it gets our device, not the child
device, and so calls the correct colour mapping procs.