PageSpotColors was getting handled by devn_get_params but not devn_put_params
For additive devices devn_put_params was not handling PageSpotColors but
devn_get_params was.  This mismatch would lead to an error for the psdrgb
device, which is an additive device built upon the DeviceN device.  Additive
devices do not handle apot colors at this time so the fix was to make
the handling of PageSpotColors in devn_get_params dependent upon the device
polarity like devn_put_params does.