Bug 692815: Fonts resourcestatus and native fonts
The core of the resourcestatus implementation for the Fonts category would
check the content of Fontmap, but would not trigger the native font scanning
to occur. So two invokations of resourcestatus for Font could result in
different results depending on whether, in the intervening period, another
operation had triggered the native font scan.

This could result in unexpected behaviour from the PDF interpreter where a
"synthetic" substitute could be created on the first page, and a native font
selected on the second page, for the same font name.

We'll now try to "find" the font for resourcestatus, if we fail, we'll try to
trigger the native font scan, and retry.

Further, I had missed the fact that this code used the Fontmap dictionary in my
previous commit, so it now accounts for the new NativeFontmap dict.

No cluster differences.