Fix for bug 695863 overprint while pattern color is active
If a pattern color space is present when a graphics state swap occurs, and
the overprint changes from on to off in the graphics state the overprint
compositor is not disabled.  In Bug 695863 a DeviceN image fill follows such
a case and the fill is improperly performed with overprint enabled. With this
commit we disable the overprint compositor if overprint is off.  If overprint
is not off, it is not clear how we should proceed, as the color space is
a factor in that set up and here the color space is a pattern.