Recursive mutex emulation: fix order of operations
We have to invalidate the self_id value *before* we unlock the mutex, as the
mutex implicitly protects the self_id (and lcount) value.

Also add asserts so we'll abort if the self_id and/or lcount get out of whack.

Finally, add a missing default value for RECURSIVE_MUTEXATTR in the ghostpdl
configure.

No cluster differences