Bug 695872: handle real/integer mixup in image Decode dict
In PDF we now pass the Height value from the image dict into the parameters
for the filter to access (specifically, the DCTDecodeFilter), and this will
cause an error if the Height value is, erroneously, a real rather than integer
value (the Postscript image operator appears to already handle that error,
at least in certain cases, gracefully).

So the filter code does not throw an error, round and convert the value to an
integer before storing it in the filter parameters dictionary.

No cluster differences.