Handle an invalid cidfmap entry in the CIDFont subst code
Certain types of invalid entry in the cidfmap can mean that resourcestatus
returns "true" but a subsequent "findresource" will fail. There is already
validation code to strip out invalid entries, but as an extra safety measure,
call the findresource in a stopped context, so we can recover, and continue
with the next substitution fallback stage.

No cluster differences.