PDF Interpreter - ignore XRefStm entries which poitn beyond end of file
Bug #695883 "Error: /invalidaccess in --run--"

The file is pretty well broken, it claims to be a hybrid (PDF 1.5 or better
readable by PDF 1.4 consumer, it uses xref streams) and as such its trailer
dictionary contains an XRefStm key whose value is supposed to point to the
PDF 1.5 xref stream. It doesn't, in fact it points past the end of the file
and so is clearly bogus.

This commit tests to ensure the XRefStm value is inside the file before
attempting to read it.

No differences expected.