PDF Interpreter - When resolving type 6 halftones, preserve the underlying PDFfile position
Bug #695886 "Error reading a content stream, missing objects"

Type 6 halftones in PDF are stream based, so when we first use a graphics
state which includes a type 6 halftone, we use resolvestream to read the
halftone.

However resolvestream does not preserve the current file position of PDFfile
and resolveht6 wasn't taking steps to save and restore it. Often this would
not be harmful but if we are already reading a stream (eg a page content
stream) and the stream was large enough not to be entirely buffered then when
we ran out of data we would return to 'PDFfile' to read another buffer from
the content stream. The fact that PDFfile had been repositioned would cause
us to read the wrong data (always, in fact, the 'endstream' from the halftone
stream).

This commit retrieves the current position of PDFfile and restores it after
resolving the stream.

No differences expected