Fix 695888 - GhostPCL locks up.
The flag to indicate a command is not implemented (e_Unimplemented) was
set to a positive integer which the GL/2 parser uses to tell the pcl
parser more PCL is expected.  This isn't a problem unless the job
terminates unexpectedly with an unimplemented GL/2 command leaving the
PCL parser poised to read more data.  Without changing the
parsers we simply change the unimplemented flag to 0 and the command
will be properly ignored, like range errors (e_Range).

Unfortunately that leaves us with the same code for e_Range and
e_Unimplemented which may be an inconvenience when debugging, but both
codes demand the same behavior of the parsers so we prefer this change
to trying to find all the places in the parser where e_Unimplemented
should be "plugged in".