Fix bug 695867, incorrect cursor position.
The current user space position was not being maintained by the HPGL IR
command.  This is fairly rare and we haven't seen it in our tests,
HPGL's SC (scale), IP (Input P1 and P2) and IR (Input Relative P1 and
P2) are used to set up scaling for the page and most jobs usually will
provide an initial user space coordinate after modifying the CTM and not
depend on the location that was current when the scaling was applied.

There is more work to be done in this area, the IP command needs similar
updating as provided here for IR, and other attributes like anchor
corner and carriage return position should be looked at for side effects
resulting from IP, IR or SC.