PDF Inerpreter - work around illegally nested BT blocks
Bug #695897 "Text missing reading PDF file"

There are actually 2 problems with this file, this is the first fix. The
file contains nested text blocks, quite a few of them, in direct contravention
of the PDF specification.

We already have code to detect this, but it relies on a specific key being
present in the current dictionary to detect the condition. Sadly this file
executes the nested blocks inside a q/Q pair as well, and the way the PDF
interpreter deals with a q means that we start a new, empty, dictionary
which neatly prevents us detecting this.

The nasty hack here introduces a new key 'qTextSaveMatrix' in the dictionary
created by a 'q' operator, *if* the current dictionary has a TextSaveMatrix
key present. We then add detection of the qTextSaveMatrix key to the BT
operator in exactly the same way as the existing TextSaveMatrx.

Attempts to have the TextSaveMatrix copied and used, so that the CTM would
follow the nesting simply didn't work, leading to the key being present
when it shouldn't be, or not present when it should, or having the wrong
value. The only 'correct' way to deal with this illegal file is to separate
the Text Matrix out from the CTM and track changes to it as required. This
would mean rewriting large parts of the PDF interpreter which I don;t think
is reasonable to cope with an invalid file.

The second problem will be the subject of another commit.

No differences expected.