PDF Interpreter - work around non-text operations inside a text block
Bug #695897 "Text missing reading PDF file"

In addition to nesting text block inside one another, the test file also
executes a 'Do' (draw image) operator inside a text block. This is also
illegal, except in the context of type character descriptions, or pattern
fills of text, or similar.

This commit slightly modifies the previous one (7bc2078) by copying the
TextSaveMatrix, instead of simply referencing it, when we execute a gsave
inside a text block. We can then safely modify the copied array whenever
a 'cm' operation takes place (just as we already did for TextSaveMatrix),
finally we can check for TextSaveMatrix or qTextSaveMatrix being present
when we execute a non-text operation. If this occurs (so far only Do tests
this) then we save the graphics state (so we don't disturb any following text)
return to the 'TextSaveMatrix', perform the operation and then restore the
graphics state.

It should be emphasised that this file is badly broken, both nested text blocks
and non-text operations inside text blocks are specifically forbidden in the
PDF reference. So the warning messages emitted when this occurs have been
extended slightly.

No differences expected