PDF interpreter - attempt to recover from broken, linearised, PDF file
Bug #695619 "mupdf cannot load page with a special PDF file"

The 'special' PDF file is in fact a totally broken PDF file, caused by
concatenating a regular PDF file with a linearised PDF file. We detect the
fact that the xref is not where the Linearised file says it is, and attempt
to recover, but fail because the trailer dictionary we find is for the
linearised file, and so does not contain a Root object. Without that we
cannot decipher the file.

We could attempt to find a 'Linearised' key, but instead I've opted to
search backwards through the file looking for a previous trailer dictionary
until we find one that has a Root key, or fail altogether to locate one. If
we find eon we proceed as before, if we fail to find a trailer with a Root
key we fail altogether, there's not much else we can do.

Note that depending how ones chooses to 'recover' from this broken PDF file
could potentially determine *which* of the 2 PDF files present actually
gets interpreted. We match Acrobat in this case, I won't guarantee we will
do so in every case.

No differences expected.