eps2write - Use an improved approach to getting the BBox of filled paths
Bug #695910 "Error reading a content stream" when using device eps2write by GhostScript 9.16 	"

The eps2write device, unlike other devices, needs a 'page' bounding box and
so often 'renders' content where the other high level devices don't. In
general this works because we degenerate to low level operations, which are
then filtered by the eps2write device and used to calculate the BBox.

Unfortunately for pattern fills this doesn't work, as we check the pattern
tile boundaries, and because we aren't really rendering these are 0. This
causes an 'unregistered' error.

This commit alters the way we calculate filled path BBoxes, instead of
rendering to low level objects, we take advantage of the code to calculate
the bounding box of a path, intersect that with the current clip path's
outer bounding box, and use that as the BBox of the path. Because we only
case about maxima and minima this is sufficient, we don't have to worry
about all the nooks and crannies potentially present in the path.

We do have to watch out for a degenerate path or bbox though, and not take
that into account.

This commit also does a couple of casts, just to eliminate a compiler warning.

No differences expected.