txtwrite - Be careful not to handle strinwidth operations as text
Bug #695914 "txtwrite command is different then Adobe save as other tool"

If we have a CIDFont with a descendant, and the descendant is not present,
and we substitute with a type 1 font, then when we do a stringwidth the
txtwrite code could throw an error (no current point) if the text was the
first element on the page.

This would then throw the font handling off, and we would get the text as
being in a type 1 font, instead of a type 0 font, which meant that the text
output was incorrect, even when a ToUnicode CMap was present.

We can deal with this easily by copying the approach from pdfwrite, if its
a stringwidth let the graphics library deal with it.

No differences in cluster, we don't test txtwrite