Fix Bug 695904, Transparency with non-identity transfer functions.
This does _not_ yet implement the arcane transfer function handling of
section 7.6.4 of the PDF 1.7 Reference Manual, but it fixes a problem
detected by customer 532 who uses a transfer function set by PS before
the PDF file runs. This was causing bands in part of the image.

The problem was that the pdf14 compositor was not applying the transfer
function for monochrome or indexed color images, and did apply it for
color images or when the halftone thresholding was done during image
processing. The transfer function was then applied during pdf14_put_image
only when the default method of passing the image to the device was
used (not when the device implemented a put_image proc, as the customer's
code does for efficiency). If color images had the transfer function
applied, then the put_image would apply it a second time.

For bands that had the pdf14 compositor skipped because transparency
was not needed, the transfer function was applied correctly, causing
differences on transitions to bands that used the pdf14 compositor.

The fix applies the transfer function in the pdf14 cmap procs and does
_not_ apply it during pdf14_put_image, which also makes it consistent
with devices that implement a put_image proc and do not have the pis
or any way to get the transfer function.