Fix for bug 695916.  Simulated overprint blending of spots colors issue
For non separation devices, we had attempted to simulate the overprint effect of spot colors
that are not CMYK values.  This was set up by default, and could be turned on/off
with -dSimulateOverprint = true/false.    Looking over the method used to
simulate the colors, it was clear that there were issues and that a completely
different approach is needed.  To do proper spot color overprinting simulation
will require buffers for each of the spot colorants.   The plan is to
achieve this through the use of the pdf14 device and setting it up in
a manner similar to how it is set up when the target device is a separation
device.  The colorants will then be collapsed to the device's real color
space when the pdf14 put image occurs.   For now, this commit disables the
blending approach, which did not work in every case, fixing the customers issue
with this file.