Big changes to pdfwrite. Previously all cos objects (composite objects) stored a pointer to the pdfwrite device, apparently only so that they would have access to the memory allocator for freeing the object.
This doesn't work if the device moves of course. So here we rework cos
objects so that they store a memory allocater instead of a device pointer
which is safe.

This did mean copying the device pointer around a few places where we had not
previously done so, so that objects could be allocated.

This commit also removes the device pointer from the 'pdf_font_cache_element'
again this was only used to free objects, we now pass the memory/device
pointer into the required routines as an argument.

I believe this completes removing the device pointers from devices, and the
code is once again stable in the face of devices being pushed and popped
even during the course of a job.