pdfwrite - Pay attention to /MediaBox pdfmarks
Previously we always emitted a /MediaBox for each page based only on the
PostScript media size, even if we had a /MediaBox /Page pdfmark.

This in itself is incorrect, but worse we would then emit the MediaBox from
the /Page pdfmark as well, leading to a dictionary with duplicate keys....

This commit will use any /MediaBox /Page pdfmark in preference to any PostScript
media settings, though it will default to those if no such pdfmark is sent.

This exhibits a difference in Bug692502.eps because this file (uniquely in
our test suite it seems) does emit a /MediaBox /Page pdfmark, which is
different to the media we use in PostScript. Previously it was incorrect,
now it is correct, or at least we honour the pdfmark now.....