PDF interpreter - additional debugging option
Added a 'stdoutprint' to go along with stderrprint, this allows us to
have the 'pdfformaterror' and other informative messages sent to stdout
instead of stderr.

When capturing stdout and stderr for debugging purposes it can be useful to
have the logging/debug messages go to the same place as the error messages
in order that they are captured properly 'interleaved'.