Tweak FAPI code for server search.
If we have multiple FAPI "servers" (i.e. font renderers) we cycle through them
testing if they can handle a particular font. To indicate it cannot, a server
will return an invalidaccess error, which we were erroneously returning to the
interpreter.

Now keep searching on an invalidaccess.

No cluster differences.