PDF Interpreter - validate and clamp arguments to Td
Bug #695950 "Incomplete rendering of PDF file"

As usual an invalid PDF file. In this case the PDF file contains an argument
to Td which is 1.5090553259294607 x 10^258. Typical implementation limits
in Appendix C of the PDF Reference limit reals to 3.403x10^38

GS returns the value as -infinity, it appears Acrobat clamps these out of range
values to 0 (at least for the Td operator)

This commit checks the arguments against 2^32 as checking against the full
range permitted for reals would be slightly more difficult, and its unlikely
that a real file would use arguments larger than this anyway. If any argument
is out of range it gets clamped to 0.

No differences expected.