Bug 695972: always use a square scaling for glyphs
We're seeing Truetype fonts whose hinting programs do not work correctly with
certain, non-square scalings. So ensure that the scale we pass to Freetype
is equal in in width and height, then apply the non-square part in the final,
output matrix (along with non-square resolution scaling, and shearing etc).

This causes quite a lot of cluster differences, due to rounding changes.