Fix error in gsapi; gsapi_run_string is broken.
gsapi_run_string calls through to gsapi_run_string_with_length
with the first argument converted from a gs_lib_ctx_t * to a void *.
gsapi_run_string_with_length then performs this conversion again.
This will lead to a SEGV.

The fix is not to do this dereference.

Wouldn't it be better if we used types here rather than void *'s?