Rejig the way we handle devices. Instead of passing the reloc/enum from the original device through the pipeline to all the child devices, we now register each device with the GC and allow them to do individual reloc/enum.
This was done because the clist maintains a pointer to the clip_path which
it does relocate, but it doesn't track it being freed. The way that the
code worked before we could end up calling the enum/reloc code in the child
devices more than once. If the clip_path had been retored away, the first
reloc would relocate the pointer to garbage, the second would try to
dereference the (now garbage) pointer and seg fault. There's no realistic
way to recode the clist, so we have to ensure the enum/reloc code only gets
called once on each GC. This ensures that happens, conveniently it also
much simplifies the enumeration/relocation code in each device.

As a result of the above, we no longer need a finalize routine, so we can
remove that complication, which means we can use the existing device macros.

However, we now face the problem that the end of job restore can free the
devices in the chain in an arbitrary order. So we need to take care to
clean up the parent/child pointers on device_finalize. We also need to
ensure that we check the child pointer before trying to use it in all the
devices, in case the child has already been freed.

Simplify the compositor code by using the 'default' compositor function.