Its possible for at least some device methods to be called even after we have started unloading and freeing the devices due to the end of job restore So here we have the default subclass code check whether there is a child device before trying to check whether it defines a method. If  there is no child then we either ignore the operation or use the gx_default method.
For the object filter especially we don't want to ignore stringwidth
(which we were doing) as later objects might be positioned based on the
return value. In addition (extra surprise) it turns out that op_show_restore
counts down the save level for a stringwidth because it *assumes* that a
gsave has taken place (this seems mad to me). Note we must pass stringwidth
operations to the child device so that the high level devices can see them.

Finally, start tidying up and implementing the object filter device.