Don't load the image compression chooser when we are going to force lossless compression
Bug #695978 "Unexpected downsampling"

As noted in the bug thread, this isn't to do with downsampling. However, in
the original email conversation I did note that there was a confusion in
two tests in the code which should have been the same.

This commit resolves that confusion, if we are forcing lossless compression
we set the count of alternate streams to 1 initially. This will prevent us
from loading the image compression chooser and hence avoid the second test
which only takes place if we have alternate streams.

Disabling the compression chooser is what gets the performance gain noted
in the bug thread. It so happens that by specifying the downsampling filter
the code took a different path, even though no downsampling was taking, and
this was causing the compression chooser to be dumped.

No differences expected.