pdfwrite - Use the Mitchell filter instead of 'bicubic' for better quality
Bug #695597 "Look at the output from the pdfwrite Bicubic image downsampler"

The Mitchell filter output is clearly of superior quality to the existing
bicubic filter. So use it instead.....

Note; when we switch to bicubic because the downsampling factor is not an
integer, we still use bicubic.

Update the documentation and change some of the defaults as well.