Have the subclass devices keep the icc_struct member as NULL instead of a copy of the child. This is because at end of job restore we can end up with the child being freed (and the icc_struct) before the subclass device.
This has some repercussions; mainly the fact that while we have a get_profile
device method, we don't have a 'set_profile' method. The icc_struct is just
stuffed straight into whichever device is on the graphics state.

Technically this could be a really bad idea, if for example the device on the
stack was a forwarding device it would not set the struct in the device it
is forwarding to. But it would forward the 'get_profile' method.