"cups"/"pwgraster" output devices: Added support for the color spaces 18 (SGray), 19 (SRGB), and 20 (Adobe RGB).
Note that this is not a full implementation in terms of color
management. The implementation is done by simply using the most
similar of the existing color spaces, simple grayscale (0) for SGray
and RGB (1) for SRGB and Adobe RGB.

This prevents Ghostscript from crashing and produce an output as close
as possible to the desired when these new color spaces are requested.

TODO: Proper color management for the SGray, SRGB, and Adobe RGB
output.