Fix bug 695929. gx_ht_construct_threshold t_level calculation was wrong.
Rewrite the threshold construction method. For num_levels < 256 the
previous "t_level_adjust", etc. were never invoked, but the logic for
the adjustment was incorrect (only worked for num_levels == 256).
This algorithm uses the same method for determining 'shade' as the
logic in gx_render_device_DeviceN which is used prior to setting up
the HT tile in the non-fast threshold code.

The 'off' is eliminated since that caused threshold values that do not
match the HT tiling code, and the init_value and threshold values now
match those generated by the HT tiling code for num_levels < 256.
Tested with 300 dpi and DITHERPPI values of 18, 24, 30, 33, 39, 52,
and 60 (the default).

Also the size of images with the fast code was wrong (dest_width)
and the phase of the pixels did not match the image code (dda_init).
This also re-enables the fast code disabled with an earlier commit.
This was tested with tests/pdf/GridDownscaleTest0.pdf, modified to
test all rotations, and at 72, 300 and 1200 dpi.

The transfer function is now pickled into the threshold array levels.
Although this causes some minor shade differences since the normal
cmap function calculates the shade from a frac (0-32760) and the
threshold has to select the shade from a byte value (0-255). The
09-47N.PS test shows this.

Also the texture screen_phase is used to adjust the thresholds when
filling the threshold buffer.

This collection of changes fixes the customer bug, also seen with
12-07B.PS and reduces the number of differences to the non-fast HT
code from 3931 to 1026. The 12-07B.PS still shows a screen angle
issue (page 3) that occurs when num_repeat > 0, but changing the logic
results in a marked increase in the number of differences.  Also, xl
tests show some phase shift in Y that remains unexplained.

All other diffs appear to be transfer function related

fixup